requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
einops==0.8.1
ffmpeg-python==0.2.0
filelock==3.18.0
flask==3.1.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
greenlet==3.2.1
huggingface-hub==0.30.2
hyperpyyaml==1.2.2
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.0
julius==0.2.7
kiwisolver==1.4.8
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip==25.0.1
primepy==1.3
propcache==0.3.1
protobuf==6.30.2
pyannote-audio==3.3.2
pyannote-core==5.0.0
pyannote-database==5.1.3
pyannote-metrics==3.2.1
pyannote-pipeline==3.0.1
pycparser==2.22
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
setuptools==80.0.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
sqlalchemy==2.0.40
sympy==1.14.0
tabulate==0.9.0
tensorboardx==2.6.2.2
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.0+cu118
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.7.0+cu118
torchmetrics==1.7.1
torchvision==0.22.0+cu118
tqdm==4.67.1
typer==0.15.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
werkzeug==3.1.3
whisper==1.1.10
yarl==1.20.0
